Modulo and List Comprehensions
Used a List Comprehension to divide all available options of the Divisors and then keep only the remainder in the list using Modulo.
If the number was 1 or 0 or a negative, it is not aPrime.
If the number had more than 2 0's in the list, it is not a Prime.
If the number had 2 0's in the list, it was a Prime.